% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mars.R
\name{plot.mars}
\alias{plot.mars}
\title{Title}
\usage{
\method{plot}{mars}(x, ...)
}
\arguments{
\item{x}{the mars object return by mars}

\item{...}{generic parameter}
}
\value{
a few plots
}
\description{
Title
}
\examples{
mm <- mars(wage ~ age,data=ISLR::Wage)
pm <- plot(mm)
}
