% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{print method for MARS (print.mars)}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{The MARS object return by mars() function.}

\item{...}{S3 generic parameter}
}
\value{
print the coefficient of mars object
}
\description{
This is the print method for MARS that print a set of useful imformations.
}
\examples{
mymars <- mars(y~.,data = marstestdata)
printmars <- print(mymars)
}
\author{
Zihe Wang, Tianle Zhong, Xiaoying Qian
}
