% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{Title}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{the mars object return by mars}

\item{...}{generic parameter}
}
\value{
the summary of mars object
}
\description{
Title
}
\examples{
mm <- mars(wage ~ age,data=ISLR::Wage)
sm <- summary(mm)
}
