% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mars.R
\name{predict.mars}
\alias{predict.mars}
\title{Title}
\usage{
\method{predict}{mars}(object, newdata, ...)
}
\arguments{
\item{object}{the mars object return by mars}

\item{newdata}{a set of new data to predict on}

\item{...}{generic parameter}
}
\value{
a list of data contain the predict values
}
\description{
Title
}
\examples{
mm <- mars(wage ~ age,data=ISLR::Wage)
pm <- predict(mm)
}
