% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mars.R
\name{print.mars}
\alias{print.mars}
\title{Title}
\usage{
\method{print}{mars}(x, ...)
}
\arguments{
\item{x}{the mars object return by mars}

\item{...}{generic parameter}
}
\value{
print the coefficient of mars object
}
\description{
Title
}
\examples{
mm <- mars(wage ~ age,data=ISLR::Wage)
pm <- print(mm)
}
