% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mars.R
\name{summary.mars}
\alias{summary.mars}
\title{summary method for MARS (summary.mars)}
\usage{
\method{summary}{mars}(object, ...)
}
\arguments{
\item{object}{The MARS object return by mars() function.}

\item{...}{S3 generic parameter}
}
\value{
A set of useful information of MARS object. Including coefficients, formula, variables, and Basis functions.
}
\description{
This is the summary method for MARS alogrithm.
}
\examples{
mymars <- mars(y~.,data = marstestdata)
summarymars <- summary(mymars)
}
\author{
Zihe Wang, Tianle Zhong, Xiaoying Qian
}
